pyproject.toml
[tool.poetry]
name = "wallet-monitor"
version = "0.1.0"
description = "Add your description here"
authors = ["David Vilela <david.vilelafreire@valory.xyz>, Jenslee Dsouza <jenslee.dsouza@valory.xyz>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4"
web3 = ">=6.0.0"
dotenv = ">=0.9.9,<0.10.0"
w3multicall = ">=0.3.1"
schedule = "^1.2.2"
python-telegram-bot = "^22.5"
